requirements.txt
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.3.0
tkinterdnd2==0.4.2